Migrate regressions from 2501-3000
