+++ spacy/tests/matcher/test_matcher_logic.py
@@ -309,6 +309,44 @@ def test_issue2464(en_vocab):
     assert len(matches) == 3
 
 
+@pytest.mark.issue(2569)
+def test_issue2569(en_tokenizer):
+    """Test that operator + is greedy."""
+    doc = en_tokenizer("It is May 15, 1993.")
+    doc.ents = [Span(doc, 2, 6, label=doc.vocab.strings["DATE"])]
+    matcher = Matcher(doc.vocab)
+    matcher.add("RULE", [[{"ENT_TYPE": "DATE", "OP": "+"}]])
+    matched = [doc[start:end] for _, start, end in matcher(doc)]
+    matched = sorted(matched, key=len, reverse=True)
+    assert len(matched) == 10
+    assert len(matched[0]) == 4
+    assert matched[0].text == "May 15, 1993"
+
+
+@pytest.mark.issue(2671)
+def test_issue2671():
+    """Ensure the correct entity ID is returned for matches with quantifiers.
+    See also #2675
+    """
+    nlp = English()
+    matcher = Matcher(nlp.vocab)
+    pattern_id = "test_pattern"
+    pattern = [
+        {"LOWER": "high"},
+        {"IS_PUNCT": True, "OP": "?"},
+        {"LOWER": "adrenaline"},
+    ]
+    matcher.add(pattern_id, [pattern])
+    doc1 = nlp("This is a high-adrenaline situation.")
+    doc2 = nlp("This is a high adrenaline situation.")
+    matches1 = matcher(doc1)
+    for match_id, start, end in matches1:
+        assert nlp.vocab.strings[match_id] == pattern_id
+    matches2 = matcher(doc2)
+    for match_id, start, end in matches2:
+        assert nlp.vocab.strings[match_id] == pattern_id
+
+
 @pytest.mark.parametrize(
     "pattern,re_pattern",
     [
