+++ spacy/tests/parser/test_ner.py
@@ -1,7 +1,9 @@
+import random
+
 import pytest
 from numpy.testing import assert_equal
-from spacy.attrs import ENT_IOB
 
+from spacy.attrs import ENT_IOB
 from spacy import util, registry
 from spacy.lang.en import English
 from spacy.lang.it import Italian
@@ -108,6 +110,28 @@ def test_issue2385():
     # maintain support for iob2 format
     tags3 = ("B-PERSON", "I-PERSON", "B-PERSON")
     assert iob_to_biluo(tags3) == ["B-PERSON", "L-PERSON", "U-PERSON"]
+
+
+@pytest.mark.issue(2800)
+def test_issue2800():
+    """Test issue that arises when too many labels are added to NER model.
+    Used to cause segfault.
+    """
+    nlp = English()
+    train_data = []
+    train_data.extend(
+        [Example.from_dict(nlp.make_doc("One sentence"), {"entities": []})]
+    )
+    entity_types = [str(i) for i in range(1000)]
+    ner = nlp.add_pipe("ner")
+    for entity_type in list(entity_types):
+        ner.add_label(entity_type)
+    optimizer = nlp.initialize()
+    for i in range(20):
+        losses = {}
+        random.shuffle(train_data)
+        for example in train_data:
+            nlp.update([example], sgd=optimizer, losses=losses, drop=0.5)
 
 
 def test_get_oracle_moves(tsys, doc, entity_annots):
